Remove unnecessary public API, move to internal
